Drop nodes at mouse location
